put paths in config file
